Added new server level middleware
Removed previous implementation of server_level middleware.

Moved all middleware (decorator or not) to the http folder.

Updated server_configs to give better typing to users.
